+++ App/Model/Post.php
@@ -22,7 +22,7 @@ class Post extends TableDataGateway
     /**
      * 定义一对多关联：一个文章有多个评论
      */
-    public ?array $hasMany = [
+    public array $hasMany = [
         [
             'tableClass' => Comment::class,
             'foreignKey' => 'post_id',